Initial release
This zip file contains the XLSM file with the VBA script, and 5 CSV workbooks for running trials. In its current state, the script is not fully functional. So, it is essentially a proof of concept for an data consolidation script. I will likely not revise it further (since it was intended for a job that I'm no longer doing), but rather, use what I learned from it to create a similar script for another automation task.